manage department and section

+++ luckySystem/src/main/java/com/example/luckySystem/dto/salary/EmployeeDailyPayrollDto.java
@@ -16,7 +16,7 @@ public class EmployeeDailyPayrollDto {
     private Long employee_daily_id;
     private Employee emp_id;
     private Long daily_pay_id;
-    private int working_hours;
+    private double working_hours;
     private double ot_amount;
     private double shift_amount;
     private double total_amount;
